Move /confirm and /deny to admin cog
